pyproject.toml
[project]
name = "meteo-to-kindle"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "httpx==0.28.1",
    "Pillow==12.0.0",
    "matplotlib==3.10.7",
    "python-dotenv==1.2.1",
]
